Suppress redundant and verbose tags from read description
It is redundant to display the read group (RG) tag in the "tag" section of
the read description, as the read group is always displayed in the basic
read information.

Other tags, like SA and MD, are very verbose and are near impossible for a
human to read.  It is cleaner to suppress them.