Reduce Graphics context creation for drawBases
The graphics context on which basepair mismatches are drawn was being allocated
once per alignment block.  For reads with many blocks (like PacBio reads), this
is very expensive and noticeably slows drawing.  To avoid repeated object creation,
create the context once per alignment and pass it down to the drawBases() method.
The object creation could be pushed even higher up the call stack to further reduce
its cost, but this handles the key bottleneck with little code change.