Simplify SAM drawAlignment() and hide small gaps
Reorganize the drawAlignment() method to be more efficient and understandable.
Rather than walking through alignment blocks and drawing the blocks, instead
walk through the alignment gaps and draw the flanking blocks.  This reorganization
makes it easier to merge blocks to skip gaps.

Use the new organization to hide small gaps that are < 3px on the screen.
This is especially useful with PacBio raw reads, which often have many gaps.