Add "consensus" mode to show only consistent mismatches
Add a "consensus" mode feature controlled by a setting that limits
the display of basepair mismatches to positions with a consistent
alternative base.  This is especially useful with third generation
sequencers that have a high single-read mismatch error rate.  The
consensus mode uses the same logic as the coverage track to determine
which mismatches to show.