Cache genome sequence to avoid multiple allocations
The reference genome sequence is used to determine whether a read
basepair is a mismatch.  Previously, the genome sequence is retrieved
for every alignment block, and sequence retrieval involves allocating
a new byte array.  This is very inefficient when there are multiple
alignments, each with many blocks.  Instead, the reference sequence
can be obtained once for each window render.