Label the size of large insertions and deletions
Introduce a "trapezoid" label for large insertions and deletions.
A purple trapezoid that points upwards indicates an insertions.  A
white trapezoid that points down indicates a deletion.  The size of
the event, in basepairs, is shown in text on the label.

This convention is particularly useful with large insertions to
differentiate between events of different scales.  It also helps
to clarify the exact size of deletions, which are hard to estimate
by eye.